requirements.txt
mysql-connector-python
pandas
numpy
watchdog

